Snapshot class is specialized
