+++ lib/event_store/entity_snapshot.rb
@@ -3,5 +3,6 @@ require 'entity_cache'
 require 'event_store/messaging'
 
 require 'event_store/entity_snapshot/activate'
+require 'event_store/entity_snapshot/entity_snapshot'
 require 'event_store/entity_snapshot/message'
 require 'event_store/entity_snapshot/store'
